requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
Django==4.0.3
dnspython==2.2.1
gunicorn==20.1.0
lxml==4.8.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.0.0
